rename "metrics" package to "metrics" package + compatibility of java 7
