+++ metrics/src/main/java/cucumber/metrics/annotation/SpeedRegulators.java
@@ -1,4 +1,4 @@
-package cucumber.metric.annotation;
+package cucumber.metrics.annotation;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
