+++ metrics/src/main/java/cucumber/metrics/core/Metered.java
@@ -1,4 +1,4 @@
-package cucumber.metric.core;
+package cucumber.metrics.core;
 
 /**
  * An object which maintains mean and exponentially-weighted rate.
