+++ metrics/src/main/java/cucumber/metrics/regulator/injector/CucumberMetricsInjectorSource.java
@@ -1,11 +1,11 @@
-package cucumber.metric.regulator.injector;
+package cucumber.metrics.regulator.injector;
 
 import com.google.inject.Guice;
 import com.google.inject.Injector;
 import com.google.inject.Stage;
 
 import cucumber.api.guice.CucumberModules;
-import cucumber.metric.regulator.module.SpeedRegulatorModule;
+import cucumber.metrics.regulator.module.SpeedRegulatorModule;
 import cucumber.runtime.java.guice.InjectorSource;
 
 public class CucumberMetricsInjectorSource implements InjectorSource {
