+++ metrics/src/test/resources/cucumber.properties
@@ -1,1 +1,1 @@
-guice.injector-source=cucumber.metric.regulator.injector.CucumberMetricsInjectorSource
+guice.injector-source=cucumber.metrics.regulator.injector.CucumberMetricsInjectorSource
